+++ game/game_libs/cl_dlls/ammo_secondary.cpp
@@ -64,7 +64,7 @@ int CHudAmmoSecondary::Draw(float)
 	UnpackRGB(r, g, b, RGB_YELLOWISH);
 	a = (int)Max(MIN_ALPHA, m_fFade);
 	if ( m_fFade > 0 )
-		m_fFade -= (gHUD.m_flTimeDelta * 20);  // slowly lower alpha to fade out icons
+		m_fFade -= static_cast<float>((gHUD.m_flTimeDelta * 20));  // slowly lower alpha to fade out icons
 	ScaleColors(r, g, b, a);
 
 	AmmoWidth = gHUD.GetSpriteRect(gHUD.m_HUD_number_0).right - gHUD.GetSpriteRect(gHUD.m_HUD_number_0).left;
